requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
pytz==2018.3
requests==2.20.0
urllib3==1.24.2
djangorestframework==3.7.7
